Use Ember services
