--- app/application/controller.js
@@ -1,9 +1,0 @@
-import Ember from 'ember';
-
-export default Ember.Controller.extend({
-
-  selectedWorker: '',
-  jobs: [],
-  selectedRows: [],
-
-});
