--- app/components/job-row/template.hbs
@@ -1,5 +1,0 @@
-<tr {{action 'select' job}}>
-  <td>{{job.assignedTo}}</td>
-  <td>{{job.location}}</td>
-  <td>{{job.status}}</td>
-</tr>
