--- app/components/worker-row/template.hbs
@@ -1,5 +1,0 @@
-<tr class={{active}} {{action "select" worker.username}}>
-  <td>{{worker.username}}</td>
-  <td>{{worker.isAvailable}}</td>
-  <td>{{worker.jobCount}}</td>
-</tr>
