+++ app/worker/service.js
@@ -1,0 +1,9 @@
+import Ember from 'ember';
+
+export default Ember.Service.extend({
+
+  workers: [],
+  selectedWorker: null,
+
+  hasSelectedWorker: Ember.computed.notEmpty('selectedWorker'),
+});
