--- tests/integration/components/job-row/component-test.js
@@ -1,13 +1,0 @@
-import { moduleForComponent, test } from 'ember-qunit';
-import hbs from 'htmlbars-inline-precompile';
-
-moduleForComponent('job-row', 'Integration | Component | job row', {
-  integration: true
-});
-
-test('select action fires', function(assert) {
-
-  this.render(hbs`{{job-row job=job select="selectJobAction"}}`);
-  assert.ok(true);
-
-});
