--- tests/integration/components/worker-row/component-test.js
@@ -1,11 +1,0 @@
-import { moduleForComponent, test } from 'ember-qunit';
-//import hbs from 'htmlbars-inline-precompile';
-
-moduleForComponent('worker-row', 'Integration | Component | worker row', {
-  integration: true
-});
-
-test('it renders', function(assert) {
-
-  assert.ok(true);
-});
